Se rompio todo

+++ integrador3/src/main/java/com/tp/integrador3/repository/EstudianteRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface EstudianteRepository extends JpaRepository <Estudiante, Long> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
 
     List<Estudiante> findAll(Sort sort);
